feat(broker): 编写 broker 存消息部分

+++ ccx-mq-remoting/src/main/java/com/ccx/mq/remoting/protocol/consts/CommandCode.java
@@ -43,7 +43,7 @@ public enum CommandCode {
      */
     private final Class<?> responseClass;
 
-    public static CommandCode fromCode(byte value) {
+    public static CommandCode fromCode(int value) {
         for (CommandCode commandCode : CommandCode.values()) {
             if (commandCode.getCode() == value) {
                 return commandCode;
